Split HTML tool-card renderer tests
